# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de bancos de dados.",
        "title": "Database Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/api/",
    "paths": {
...
        "/database": {
            "get": {
                "description": "Retorna todas os bancos de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "Lista de banco de dados",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Altera um banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "Altera banco de dados",
                "parameters": [
                    {
                        "description": "Banco de Dados",
                        "name": "database",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DatabaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Cria um banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "Cria banco de dados",
                "parameters": [
                    {
                        "description": "Banco de Dados",
                        "name": "database",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DatabaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/database/select": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "Seleciona um banco de dados",
                "parameters": [
                    {
                        "description": "Banco de Dados",
                        "name": "database",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DatabaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/table": {
            "get": {
                "description": "Retorna todas as tabelas do banco selecionado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Lista de tabelas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Cria tabela",
                "parameters": [
                    {
                        "description": "Tabela nova",
                        "name": "tabela",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Table"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
...
        "/table/item": {
            "put": {
                "description": "Altera um item em uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Altera item",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um item em uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Cria item",
                "parameters": [
                    {
                        "description": "Item novo",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um item de uma tabela",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Remove item",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TableData"
                        }
                    }
                }
            }
        },
        "/table/item/{table}": {
            "get": {
                "description": "Retorna uma lista com as colunas e uma com os itens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Lista colunas e itens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tabela",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TableData"
                        }
                    }
                }
            }
        },
        "/table/{name}": {
            "delete": {
                "description": "Remove o banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Remove o banco de dados",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da tabela",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/table/{oldName}/{newName}": {
            "put": {
                "description": "Altera uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Altera tabela",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome atual",
                        "name": "oldName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nome novo",
                        "name": "newName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/table/{table}": {
            "delete": {
                "description": "Remove uma tabela",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Remove tabela",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da tabela",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Column": {
            "type": "object",
            "properties": {
                "autoIncrement": {
                    "type": "boolean",
                    "example": false
                },
                "defaultValue": {
                    "type": "string",
                    "example": ""
                },
                "length": {
                    "type": "integer",
                    "example": 10
                },
                "name": {
                    "type": "string",
                    "example": "teste"
                },
                "nullable": {
                    "type": "boolean",
                    "example": true
                },
                "primary": {
                    "type": "boolean",
                    "example": false
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "unique": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.DatabaseRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "my_system"
                },
                "newName": {
                    "type": "string",
                    "example": "db_teste1"
                },
                "oldName": {
                    "type": "string",
                    "example": "db_teste"
                }
            }
        },
        "models.ItemRequest": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "table": {
                    "type": "string"
                },
                "values": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "models.Table": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Column"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "teste"
                }
            }
        },
        "models.TableData": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        },
        "respond.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}